backend/requirements.txt
fastapi
uvicorn[standard]
gunicorn
pydantic
pydantic-settings
python-multipart
orjson
anyio
starlette
pillow
numpy